:recycle: Refactor: reorganize project structure
