+++ index.html
@@ -7,91 +7,6 @@
     <meta name="google-adsense-account" content="ca-pub-2515847504885873">
     <title>Schoolish</title>
 
-    <style>
-      @import url("https://fonts.googleapis.com/css2?family=Poppins");
-
-      * {
-        padding: 0;
-        margin: 0;
-        box-sizing: border-box;
-        font-family: "Poppins", sans-serif;
-      }
-
-      :root {
-        --mainColor: #7b00c3;
-        --mainColorLight: #5767aa;
-        --secondaryColor: #c99de3;
-        --textColor: #eee;
-      }
-
-      .arrow {
-        height: 12px;
-        width: 12px;
-        border-bottom: 4px solid;
-        border-right: 4px solid;
-        border-bottom-right-radius: 20%;
-      }
-
-      .arrow-right {
-        rotate: -45deg;
-      }
-
-      #alertMessageBox {
-        padding: 15px;
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-        border-radius: 5px;
-        box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
-      }
-
-      /* Loader */
-      .loading {
-        width: 100vw;
-        height: 100vh;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        background-color: #7b00c3;
-      }
-      .loader {
-        --s: 50px;
-        --_d: calc(0.353 * var(--s));
-
-        height: calc(var(--s) + var(--_d));
-        aspect-ratio: 1;
-        display: grid;
-      }
-      .loader:before {
-        content: "";
-        height: 100%;
-        margin: auto 0;
-        clip-path: polygon(
-          var(--_d) 0,
-          100% 0,
-          100% calc(100% - var(--_d)),
-          calc(100% - var(--_d)) 100%,
-          0 100%,
-          0 var(--_d)
-        );
-        background: conic-gradient(
-          from -90deg at var(--s) var(--_d),
-          #fff 135deg,
-          #666 0 270deg,
-          #aaa 0
-        );
-        animation: l3 0.8s infinite alternate;
-      }
-      @keyframes l3 {
-        100% {
-          height: 40%;
-        }
-      }
-      .displayNone {
-        display: none;
-      }
-    </style>
 
     <!-- JavaScript -->
     <script
@@ -101,19 +16,17 @@
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js"></script>
 
-    <!-- Stylesheet -->
+    <!-- Icon -->
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <link
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
     />
   </head>
+
   <body>
     <div id="root">
-      <div class="loading">
-        <div class="loader"></div>
-      </div>
+      <script type="module" src="/src/main.tsx"></script>
     </div>
-    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
