+++ package.json
@@ -8,7 +8,7 @@
   "repository": "https://github.com/MigPlayer6032/schoolish",
   "author": "Miguel Tibincoski",
   "scripts": {
-    "dev": "cls && vite",
+    "dev": "vite",
     "build": "tsc -b && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
