+++ src/main.tsx
@@ -1,31 +1,21 @@
-import { Suspense } from "react";
-import * as ReactDOM from "react-dom/client";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { SpeedInsights } from "@vercel/speed-insights/react";
+import { StrictMode } from "react";
+// import creatRoot 
+import { createRoot } from "react-dom/client";
 
-import Home from "./pages/Home";
-import Signup from "./pages/Signup";
-import Login from "./pages/Login";
-import Dashboard from "./pages/Dashboard";
-
+// import global css and APP and move coponents to main
+import App from "./App";
+import './styles/global.css'
 import "./styles/firebaseEmulatorWarning.css";
-import Loading from "./components/Loading";
 
-import "./services/i18n";
+import "./services/i18n"; // I dont know what this is :/
 
-ReactDOM.createRoot(
-  document.getElementById("root") ?? document.createElement("div")
-).render(
-  <Suspense fallback={<Loading />}>
-    <BrowserRouter>
-      <SpeedInsights />
-
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/signup" element={<Signup />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/dashboard" element={<Dashboard />} />
-      </Routes>
-    </BrowserRouter>
-  </Suspense>
+// removing create element
+createRoot(
+  document.getElementById("root")!).render(
+    <StrictMode>
+      <div className="loading">
+        <div className="loader"></div>
+        <App />
+      </div>
+    </StrictMode>
 );
