+++ src/styles/global.css
@@ -1,0 +1,77 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+  font-family: "Poppins", sans-serif;
+}
+:root {
+  --mainColor: #7b00c3;
+  --mainColorLight: #5767aa;
+  --secondaryColor: #c99de3;
+  --textColor: #eee;
+}
+.arrow {
+  height: 12px;
+  width: 12px;
+  border-bottom: 4px solid;
+  border-right: 4px solid;
+  border-bottom-right-radius: 20%;
+}
+.arrow-right {
+  rotate: -45deg;
+}
+#alertMessageBox {
+  padding: 15px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  border-radius: 5px;
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
+}
+/* Loader */
+.loading {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #7b00c3;
+}
+.loader {
+  --s: 50px;
+  --_d: calc(0.353 * var(--s));
+  height: calc(var(--s) + var(--_d));
+  aspect-ratio: 1;
+  display: grid;
+}
+.loader:before {
+  content: "";
+  height: 100%;
+  margin: auto 0;
+  clip-path: polygon(
+    var(--_d) 0,
+    100% 0,
+    100% calc(100% - var(--_d)),
+    calc(100% - var(--_d)) 100%,
+    0 100%,
+    0 var(--_d)
+  );
+  background: conic-gradient(
+    from -90deg at var(--s) var(--_d),
+    #fff 135deg,
+    #666 0 270deg,
+    #aaa 0
+  );
+  animation: l3 0.8s infinite alternate;
+}
+@keyframes l3 {
+  100% {
+    height: 40%;
+  }
+}
+.displayNone {
+  display: none;
+}
